pyproject.toml
[project]
name = "langgraph-langchain"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "arxiv>=2.2.0",
    "bs4>=0.0.2",
    "chromadb>=1.0.20",
    "ipykernel>=6.30.1",
    "langchain>=0.3.27",
    "langchain-community>=0.3.29",
    "langchain-huggingface>=0.3.1",
    "langchain-openai>=0.3.32",
    "langchain-text-splitters>=0.3.11",
    "pymupdf>=1.26.4",
    "pypdf>=6.0.0",
    "python-dotenv>=1.1.1",
    "sentence-transformers>=5.1.0",
]
